# api/v1alpha1/openapi.yaml
openapi: 3.0.1
info:
  title: Flight Control API
  version: v1alpha1
  description: |
    [Flight Control](https://flightctl.io) is a service for declarative management of fleets of edge devices and their workloads.
  contact:
    name: The Flight Control Team
    url: https://flightctl.io
    email: team@flightctl.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
...
  /api/v1/organizations/{orgID}/certificatesigningrequests:
    get:
      tags:
        - certificatesigningrequest
      description: List CertificateSigningRequest resources.
      operationId: listCertificateSigningRequests
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: continue
          in: query
          description: An optional parameter to query more results from the server. The value of the paramter must match the value of the 'continue' field in the previous list response.
          required: false
          schema:
            type: string
        - name: labelSelector
          in: query
          description: A selector to restrict the list of returned objects by their labels. Defaults to everything.
          schema:
            type: string
        - name: fieldSelector
          in: query
          description: A selector to restrict the list of returned objects by their fields, supporting operators like '=', '==', and '!=' (e.g., "key1=value1,key2!=value2").
          schema:
            type: string
        - name: limit
          in: query
          description: The maximum number of results returned in the list response. The server will set the 'continue' field in the list response if more results exist. The continue value may then be specified as parameter in a subsequent query.
          required: false
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateSigningRequestList'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    post:
      tags:
        - certificatesigningrequest
      description: Create a CertificateSigningRequest resource.
      operationId: createCertificateSigningRequest
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateSigningRequest'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateSigningRequest'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: StatusConflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /api/v1/organizations/{orgID}/certificatesigningrequests/{name}:
    patch:
      tags:
        - certificatesigningrequest
      description: Patch a CertificateSigningRequest resource.
      operationId: patchCertificateSigningRequest
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the CertificateSigningRequest resource to patch.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/PatchRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateSigningRequest'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    get:
      tags:
        - certificatesigningrequest
      description: read the specified certificateSigningRequest
      operationId: getCertificateSigningRequest
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the CertificateSigningRequest resource to get.
          required: true
          schema:
            type: string
            uniqueItems: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateSigningRequest'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    put:
      tags:
        - certificatesigningrequest
      description: replace the specified CertificateSigningRequest
      operationId: replaceCertificateSigningRequest
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the CertificateSigningRequest resource to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateSigningRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateSigningRequest'
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateSigningRequest'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    delete:
      tags:
        - certificatesigningrequest
      description: delete a Certificate Signing Request
      operationId: deleteCertificateSigningRequest
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the CertificateSigningRequest resource to delete.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /api/v1/organizations/{orgID}/certificatesigningrequests/{name}/approval:
    put:
      tags:
        - certificatesigningrequest
      description: Approve or deny a CertificateSigningRequest.
      operationId: updateCertificateSigningRequestApproval
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the CertificateSigningRequest to approve or deny.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateSigningRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateSigningRequest'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: StatusConflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
components:
  schemas:
...
    PatchRequest:
      type: array
      items:
        type: object
        additionalProperties: false
        required:
          - op
          - path
        properties:
          path:
            description: A JSON Pointer path.
            type: string
          value:
            description: The value to add or replace.
          op:
            description: The operation to perform.
            type: string
            enum:
              - add
              - replace
              - remove
...
    ListMeta:
      type: object
      properties:
        continue:
          type: string
          description: May be set if the user set a limit on the number of items returned, and indicates that the server has more data available. The value is opaque and may be used to issue another request to the endpoint that served this list to retrieve the next set of available objects. Continuing a consistent list may not be possible if the server configuration has changed or more than a few minutes have passed. The resourceVersion field returned when using this continue value will be identical to the value in the first response, unless you have received this token from an error message.
        remainingItemCount:
          type: integer
          description: The number of subsequent items in the list which are not included in this list response. If the list request contained label or field selectors, then the number of remaining items is unknown and the field will be left unset and omitted during serialization. If the list is complete (either because it is not chunking or because this is the last chunk), then there are no more remaining items and this field will be left unset and omitted during serialization. Servers older than v1.15 do not set this field. The intended use of the remainingItemCount is *estimating* the size of a collection. Clients should not rely on the remainingItemCount to be set or to be exact.
          format: int64
      description: ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.
    ObjectMeta:
      type: object
      properties:
        creationTimestamp:
          type: string
          description: The time the object was created.
          format: date-time
        deletionTimestamp:
          type: string
          description: The time the object will be deleted.
          format: date-time
        name:
          type: string
          description: The name of the object.
        labels:
          type: object
          description: Map of string keys and values that can be used to organize and categorize (scope and select) objects.
          additionalProperties:
            type: string
        generation:
          type: integer
          description: A sequence number representing a specific generation of the desired state. Populated by the system. Read-only.
          format: int64
        owner:
          type: string
          description: A resource that owns this resource, in "kind/name" format.
        annotations:
          type: object
          additionalProperties:
            type: string
          description: Properties set by the service.
        resourceVersion:
          type: string
          description: An opaque string that identifies the server's internal version of an object.
      description: ObjectMeta is metadata that all persisted resources must have, which includes all objects users must create.
...
    Status:
      type: object
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        code:
          type: integer
          format: int32
          description: Suggested HTTP return code for this status, 0 if not set.
        message:
          type: string
          description: A human-readable description of the status of this operation.
        reason:
          type: string
          description: A machine-readable description of why this operation is in the "Failure" status. If this value is empty there is no information available. A Reason clarifies an HTTP status code but does not override it.
        status:
          type: string
          description: 'Status of the operation. One of: "Success" or "Failure". More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status.'
      description: Status is a return value for calls that don't return other objects.
      required:
        - apiVersion
        - kind
        - status
        - code
        - reason
        - message
    Condition:
      required:
        - type
        - status
        - lastTransitionTime
        - reason
        - message
      type: object
      properties:
        type:
          $ref: '#/components/schemas/ConditionType'
        status:
          $ref: '#/components/schemas/ConditionStatus'
        observedGeneration:
          type: integer
          format: int64
          description: The .metadata.generation that the condition was set based upon.
        lastTransitionTime:
          type: string
          format: date-time
          description: The last time the condition transitioned from one status to another.
        message:
          type: string
          description: Human readable message indicating details about last transition.
        reason:
          type: string
          description: A (brief) reason for the condition's last transition.
      description: Condition contains details for one aspect of the current state of this API Resource.
    ConditionType:
      type: string
      description: Type of condition in CamelCase.
      enum:
      - 'Approved'              # EnrollmentRequest
      - 'Approved'              # CertificateSigningRequest
      - 'Denied'                # CertificateSigningRequest
      - 'Failed'                # CertificateSigningRequest
      - 'Accessible'            # Repository
      - 'Accessible'            # ResourceSync
      - 'ResourceParsed'        # ResourceSync
      - 'Synced'                # ResourceSync
      - 'OverlappingSelectors'  # Fleet
      - 'Valid'                 # Fleet
      - 'RolloutInProgress'     # Fleet
      - 'Updating'              # Device
      - 'SpecValid'             # Device (service condition)
      - 'MultipleOwners'        # Device (service condition)
      - 'DeviceDecommissioning' # Device
      x-enum-varnames:
      - EnrollmentRequestApproved
      - CertificateSigningRequestApproved
      - CertificateSigningRequestDenied
      - CertificateSigningRequestFailed
      - RepositoryAccessible
      - ResourceSyncAccessible
      - ResourceSyncResourceParsed
      - ResourceSyncSynced
      - FleetOverlappingSelectors
      - FleetValid
      - FleetRolloutInProgress
      - DeviceUpdating
      - DeviceSpecValid
      - DeviceMultipleOwners
      - DeviceDecommissioning
    ConditionStatus:
      type: string
      description: Status of the condition, one of True, False, Unknown.
      enum:
      - 'True'
      - 'False'
      - 'Unknown'
      x-enum-varnames:
      - 'ConditionStatusTrue'
      - 'ConditionStatusFalse'
      - 'ConditionStatusUnknown'
    CertificateSigningRequest:
      type: object
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        metadata:
          $ref: '#/components/schemas/ObjectMeta'
        spec:
          $ref: '#/components/schemas/CertificateSigningRequestSpec'
        status:
          $ref: '#/components/schemas/CertificateSigningRequestStatus'
      required:
        - apiVersion
        - kind
        - metadata
        - spec
      description: 'CertificateSigningRequest represents a request for a signed certificate from the CA.'
    CertificateSigningRequestList:
      type: object
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        metadata:
          $ref: '#/components/schemas/ListMeta'
        items:
          type: array
          description: 'List of CertificateSigningRequest.'
          items:
            $ref: '#/components/schemas/CertificateSigningRequest'
      required:
        - apiVersion
        - kind
        - metadata
        - items
      description: CertificateSigningRequestList is a list of CertificateSigningRequest.
    CertificateSigningRequestSpec:
      description: Wrapper around a user-created CSR, modeled on kubernetes io.k8s.api.certificates.v1.CertificateSigningRequestSpec.
      properties:
        expirationSeconds:
          description: Requested duration of validity for the certificate.
          format: int32
          type: integer
        extra:
          additionalProperties:
            items:
              type: string
            type: array
          description: Extra attributes of the user that created the CSR, populated by the API server on creation and immutable.
          type: object
        request:
          description: The base64-encoded PEM-encoded PKCS#10 CSR. Matches the spec.request field in a kubernetes CertificateSigningRequest resource.
          format: byte
          type: string
        signerName:
          description: Indicates the requested signer, and is a qualified name.
          type: string
        uid:
          description: UID of the user that created the CSR, populated by the API server on creation and immutable.
          type: string
        usages:
          description: Usages specifies a set of key usages requested in the issued certificate.
          items:
            type: string
          type: array
        username:
          description: Name of the user that created the CSR, populated by the API server on creation and immutable.
          type: string
      required:
        - request
        - signerName
      type: object
    CertificateSigningRequestStatus:
      description: Indicates approval/denial/failure status of the CSR, and contains the issued certificate if any exists.
      properties:
        certificate:
          description: The issued signed certificate, immutable once populated.
          format: byte
          type: string
        conditions:
          type: array
          description: Conditions applied to the request. Known conditions are Approved, Denied, and Failed.
          items:
            $ref: '#/components/schemas/Condition'
      required:
        - conditions
      type: object
...